Add Bubble and Waypoints components
